Run long integration tests parallel
mocha-parallel-tests for src/integration.long/*.test.ts takes 2m37.844s
while mocha takes 6m52.098s.
